requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.6
langchain-openai==0.1.13
langchain-community==0.2.6
langgraph==0.1.14
chromadb==0.5.3
sentence-transformers==3.0.1
pypdf2==3.0.1
pydantic==2.7.4
python-multipart==0.0.9
sqlalchemy==2.0.31
aiofiles==23.2.1
httpx==0.27.0
streamlit==1.36.0
requests==2.32.3
python-dotenv==1.0.1
groq==0.9.0
